defect #1060041 Add dismiss in context menu for comments in my work

+++ octane-eclipse-plugin/src/com/hpe/octane/ideplugins/eclipse/ui/mywork/job/DismissItemJob.java
@@ -18,24 +18,34 @@ import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
 
 import com.hpe.adm.nga.sdk.model.EntityModel;
+import com.hpe.adm.octane.ideplugins.services.CommentService;
+import com.hpe.adm.octane.ideplugins.services.filtering.Entity;
 import com.hpe.adm.octane.ideplugins.services.mywork.MyWorkService;
 import com.hpe.octane.ideplugins.eclipse.Activator;
 
 public class DismissItemJob extends Job {
 
     MyWorkService myWorkService = Activator.getInstance(MyWorkService.class);
+    private CommentService commentService = Activator.getInstance(CommentService.class);
     EntityModel entityModel;
+    Entity entityType;
     private boolean wasRemoved = false;
 
-    public DismissItemJob(String name, EntityModel entityModel) {
+    public DismissItemJob(String name, EntityModel entityModel, Entity entityType) {
         super(name);
         this.entityModel = entityModel;
+        this.entityType = entityType;
     }
 
     @Override
     protected IStatus run(IProgressMonitor monitor) {
         monitor.beginTask(getName(), IProgressMonitor.UNKNOWN);
-        wasRemoved = myWorkService.removeFromMyWork(entityModel);
+        if (entityType == Entity.COMMENT) {
+        	wasRemoved = commentService.dismissComment(entityModel);
+        } 
+        else {
+        	wasRemoved = myWorkService.removeFromMyWork(entityModel);
+        }
         monitor.done();
         return Status.OK_STATUS;
     }
